=str #16309 DRY FlowIteratorSpec, FlowIterableSpec, SynchronousIterableSpec
