+++ akka-stream-tests/src/test/scala/akka/stream/scaladsl/FlowIteratorSpec.scala
@@ -5,7 +5,6 @@ package akka.stream.scaladsl
 
 import scala.collection.immutable
 import scala.concurrent.duration._
-
 import akka.stream.ActorFlowMaterializer
 import akka.stream.ActorFlowMaterializerSettings
 import akka.stream.testkit.AkkaSpec
@@ -13,17 +12,139 @@ import akka.stream.testkit.StreamTestKit
 import akka.stream.testkit.StreamTestKit.OnComplete
 import akka.stream.testkit.StreamTestKit.OnError
 import akka.stream.testkit.StreamTestKit.OnNext
+import akka.stream.impl.SynchronousIterablePublisher
+import org.reactivestreams.Subscription
+import akka.testkit.TestProbe
+import org.reactivestreams.Subscriber
 
 class FlowIteratorSpec extends AbstractFlowIteratorSpec {
   override def testName = "A Flow based on an iterator producing function"
-  override def createSource[T](iterable: immutable.Iterable[T]): Source[T, Unit] =
-    Source(() ⇒ iterable.iterator)
+  override def createSource(elements: Int): Source[Int, Unit] =
+    Source(() ⇒ (1 to elements).iterator)
 }
 
 class FlowIterableSpec extends AbstractFlowIteratorSpec {
   override def testName = "A Flow based on an iterable"
-  override def createSource[T](iterable: immutable.Iterable[T]): Source[T, Unit] =
-    Source(iterable)
+  override def createSource(elements: Int): Source[Int, Unit] =
+    Source(1 to elements)
+
+  implicit def mmaterializer = super.materializer
+
+  "produce onError when iterator throws" in {
+    val iterable = new immutable.Iterable[Int] {
+      override def iterator: Iterator[Int] =
+        (1 to 3).iterator.map(x ⇒ if (x == 2) throw new IllegalStateException("not two") else x)
+    }
+    val p = Source(iterable).runWith(Sink.publisher)
+    val c = StreamTestKit.SubscriberProbe[Int]()
+    p.subscribe(c)
+    val sub = c.expectSubscription()
+    sub.request(1)
+    c.expectNext(1)
+    c.expectNoMsg(100.millis)
+    sub.request(2)
+    c.expectError.getMessage should be("not two")
+    sub.request(2)
+    c.expectNoMsg(100.millis)
+  }
+
+  "produce onError when Source construction throws" in {
+    val iterable = new immutable.Iterable[Int] {
+      override def iterator: Iterator[Int] = throw new IllegalStateException("no good iterator")
+    }
+    val p = Source(iterable).runWith(Sink.publisher)
+    val c = StreamTestKit.SubscriberProbe[Int]()
+    p.subscribe(c)
+    c.expectSubscriptionAndError().getMessage should be("no good iterator")
+    c.expectNoMsg(100.millis)
+  }
+
+  "produce onError when hasNext throws" in {
+    val iterable = new immutable.Iterable[Int] {
+      override def iterator: Iterator[Int] = new Iterator[Int] {
+        override def hasNext: Boolean = throw new IllegalStateException("no next")
+        override def next(): Int = -1
+      }
+    }
+    val p = Source(iterable).runWith(Sink.publisher)
+    val c = StreamTestKit.SubscriberProbe[Int]()
+    p.subscribe(c)
+    c.expectSubscriptionAndError().getMessage should be("no next")
+    c.expectNoMsg(100.millis)
+  }
+}
+
+class SynchronousIterableSpec extends AbstractFlowIteratorSpec {
+  override def testName = "A Flow based on small collection"
+  override def createSource(elements: Int): Source[Int, Unit] =
+    Source(SynchronousIterablePublisher(1 to elements, "range"))
+
+  "not produce after cancel from onNext" in {
+    val p = SynchronousIterablePublisher(1 to 5, "range")
+    val probe = TestProbe()
+    p.subscribe(new Subscriber[Int] {
+      var sub: Subscription = _
+      override def onError(cause: Throwable): Unit = probe.ref ! cause
+      override def onComplete(): Unit = probe.ref ! "complete"
+      override def onNext(element: Int): Unit = {
+        probe.ref ! element
+        if (element == 3) sub.cancel()
+      }
+      override def onSubscribe(subscription: Subscription): Unit = {
+        sub = subscription
+        sub.request(10)
+      }
+    })
+
+    probe.expectMsg(1)
+    probe.expectMsg(2)
+    probe.expectMsg(3)
+    probe.expectNoMsg(500.millis)
+  }
+
+  "produce onError when iterator throws" in {
+    val iterable = new immutable.Iterable[Int] {
+      override def iterator: Iterator[Int] =
+        (1 to 3).iterator.map(x ⇒ if (x == 2) throw new IllegalStateException("not two") else x)
+    }
+    val p = SynchronousIterablePublisher(iterable, "iterable")
+    val c = StreamTestKit.SubscriberProbe[Int]()
+    p.subscribe(c)
+    val sub = c.expectSubscription()
+    sub.request(1)
+    c.expectNext(1)
+    c.expectNoMsg(100.millis)
+    sub.request(2)
+    c.expectError.getMessage should be("not two")
+    sub.request(2)
+    c.expectNoMsg(100.millis)
+  }
+
+  "handle reentrant requests" in {
+    val N = 50000
+    val p = SynchronousIterablePublisher(1 to N, "range")
+    val probe = TestProbe()
+    p.subscribe(new Subscriber[Int] {
+      var sub: Subscription = _
+      override def onError(cause: Throwable): Unit = probe.ref ! cause
+      override def onComplete(): Unit = probe.ref ! "complete"
+      override def onNext(element: Int): Unit = {
+        probe.ref ! element
+        sub.request(1)
+
+      }
+      override def onSubscribe(subscription: Subscription): Unit = {
+        sub = subscription
+        sub.request(1)
+      }
+    })
+    probe.receiveN(N) should be((1 to N).toVector)
+    probe.expectMsg("complete")
+  }
+
+  "have a toString that doesn't OOME" in {
+    SynchronousIterablePublisher(1 to 3, "range").toString should be("range")
+  }
 }
 
 abstract class AbstractFlowIteratorSpec extends AkkaSpec {
@@ -31,15 +152,16 @@ abstract class AbstractFlowIteratorSpec extends AkkaSpec {
   val settings = ActorFlowMaterializerSettings(system)
     .withInputBuffer(initialSize = 2, maxSize = 2)
 
-  implicit val materializer = ActorFlowMaterializer(settings)
+  private val m = ActorFlowMaterializer(settings)
+  implicit final def materializer = m
 
   def testName: String
 
-  def createSource[T](iterable: immutable.Iterable[T]): Source[T, Unit]
+  def createSource(elements: Int): Source[Int, Unit]
 
   testName must {
     "produce elements" in {
-      val p = createSource(1 to 3).runWith(Sink.publisher)
+      val p = createSource(3).runWith(Sink.publisher)
       val c = StreamTestKit.SubscriberProbe[Int]()
       p.subscribe(c)
       val sub = c.expectSubscription()
@@ -53,7 +175,7 @@ abstract class AbstractFlowIteratorSpec extends AkkaSpec {
     }
 
     "complete empty" in {
-      val p = createSource(immutable.Iterable.empty[Int]).runWith(Sink.publisher)
+      val p = createSource(0).runWith(Sink.publisher)
       val c = StreamTestKit.SubscriberProbe[Int]()
       p.subscribe(c)
       c.expectSubscriptionAndComplete()
@@ -61,7 +183,7 @@ abstract class AbstractFlowIteratorSpec extends AkkaSpec {
     }
 
     "produce elements with multiple subscribers" in {
-      val p = createSource(1 to 3).runWith(Sink.fanoutPublisher(2, 4))
+      val p = createSource(3).runWith(Sink.fanoutPublisher(2, 4))
       val c1 = StreamTestKit.SubscriberProbe[Int]()
       val c2 = StreamTestKit.SubscriberProbe[Int]()
       p.subscribe(c1)
@@ -85,7 +207,7 @@ abstract class AbstractFlowIteratorSpec extends AkkaSpec {
     }
 
     "produce elements to later subscriber" in {
-      val p = createSource(1 to 3).runWith(Sink.fanoutPublisher(2, 4))
+      val p = createSource(3).runWith(Sink.fanoutPublisher(2, 4))
       val c1 = StreamTestKit.SubscriberProbe[Int]()
       val c2 = StreamTestKit.SubscriberProbe[Int]()
       p.subscribe(c1)
@@ -108,7 +230,7 @@ abstract class AbstractFlowIteratorSpec extends AkkaSpec {
     }
 
     "produce elements with one transformation step" in {
-      val p = createSource(1 to 3).map(_ * 2).runWith(Sink.publisher)
+      val p = createSource(3).map(_ * 2).runWith(Sink.publisher)
       val c = StreamTestKit.SubscriberProbe[Int]()
       p.subscribe(c)
       val sub = c.expectSubscription()
@@ -120,7 +242,7 @@ abstract class AbstractFlowIteratorSpec extends AkkaSpec {
     }
 
     "produce elements with two transformation steps" in {
-      val p = createSource(1 to 4).filter(_ % 2 == 0).map(_ * 2).runWith(Sink.publisher)
+      val p = createSource(4).filter(_ % 2 == 0).map(_ * 2).runWith(Sink.publisher)
       val c = StreamTestKit.SubscriberProbe[Int]()
       p.subscribe(c)
       val sub = c.expectSubscription()
@@ -131,7 +253,7 @@ abstract class AbstractFlowIteratorSpec extends AkkaSpec {
     }
 
     "not produce after cancel" in {
-      val p = createSource(1 to 3).runWith(Sink.publisher)
+      val p = createSource(3).runWith(Sink.publisher)
       val c = StreamTestKit.SubscriberProbe[Int]()
       p.subscribe(c)
       val sub = c.expectSubscription()
@@ -142,47 +264,5 @@ abstract class AbstractFlowIteratorSpec extends AkkaSpec {
       c.expectNoMsg(100.millis)
     }
 
-    "produce onError when iterator throws" in {
-      val iterable = new immutable.Iterable[Int] {
-        override def iterator: Iterator[Int] =
-          (1 to 3).iterator.map(x ⇒ if (x == 2) throw new IllegalStateException("not two") else x)
-      }
-      val p = createSource(iterable).runWith(Sink.publisher)
-      val c = StreamTestKit.SubscriberProbe[Int]()
-      p.subscribe(c)
-      val sub = c.expectSubscription()
-      sub.request(1)
-      c.expectNext(1)
-      c.expectNoMsg(100.millis)
-      sub.request(2)
-      c.expectError.getMessage should be("not two")
-      sub.request(2)
-      c.expectNoMsg(100.millis)
-    }
-
-    "produce onError when Source construction throws" in {
-      val iterable = new immutable.Iterable[Int] {
-        override def iterator: Iterator[Int] = throw new IllegalStateException("no good iterator")
-      }
-      val p = createSource(iterable).runWith(Sink.publisher)
-      val c = StreamTestKit.SubscriberProbe[Int]()
-      p.subscribe(c)
-      c.expectSubscriptionAndError().getMessage should be("no good iterator")
-      c.expectNoMsg(100.millis)
-    }
-
-    "produce onError when hasNext throws" in {
-      val iterable = new immutable.Iterable[Int] {
-        override def iterator: Iterator[Int] = new Iterator[Int] {
-          override def hasNext: Boolean = throw new IllegalStateException("no next")
-          override def next(): Int = -1
-        }
-      }
-      val p = createSource(iterable).runWith(Sink.publisher)
-      val c = StreamTestKit.SubscriberProbe[Int]()
-      p.subscribe(c)
-      c.expectSubscriptionAndError().getMessage should be("no next")
-      c.expectNoMsg(100.millis)
-    }
   }
 }
